+++ packages/babel-traverse/src/cache.js
